refactored creating/updating file
