+++ Client/src/app/services/file.service.ts
@@ -10,6 +10,25 @@ export class FileService {
 
     constructor() { }
 
+    public async CreateFile(name: string, text: string): Promise<string> {
+        let url = environment.filesUri;
+        let params = {
+            headers: {
+                "content-type": "application/json; charset=utf-8",
+                "Authorization": `Bearer ${localStorage.getItem("token")}`
+            },
+            body: JSON.stringify({
+                Name: name,
+                Text: text
+            }),
+            method: "POST"
+        }
+
+        var resp = await fetch(url, params);
+        var result = await resp.json();
+        return result.id;
+    }
+
     public async getFiles(): Promise<File[]> {
 
         let url = environment.filesUri;
